# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "paper API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/trades": {
      "post": {
        "summary": "Execute one buy/sell at the cached price",
        "tags": ["trades"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "asset_id": {
                    "type": "string",
                    "enum": [
                      "BTC",
                      "ETH",
                      "SOL",
                      "USDC",
                      "BNB",
                      "XRP",
                      "ADA",
                      "DOGE",
                      "AVAX",
                      "LINK",
                      "DOT",
                      "TON"
                    ]
                  },
                  "side": {
                    "type": "string",
                    "enum": ["buy", "sell"]
                  },
                  "usd_amount": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,8})?$"
                  },
                  "idempotency_key": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  }
                },
                "required": ["asset_id", "side", "usd_amount", "idempotency_key"]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trade": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "asset_id": {
                          "type": "string"
                        },
                        "side": {
                          "type": "string",
                          "enum": ["buy", "sell"]
                        },
                        "usd_amount": {
                          "type": "string"
                        },
                        "qty": {
                          "type": "string"
                        },
                        "price_at_execution": {
                          "type": "string"
                        },
                        "idempotency_key": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "asset_id",
                        "side",
                        "usd_amount",
                        "qty",
                        "price_at_execution",
                        "idempotency_key",
                        "created_at"
                      ],
                      "additionalProperties": false
                    },
                    "is_first_trade": {
                      "type": "boolean"
                    }
                  },
                  "required": ["trade", "is_first_trade"],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trade": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "asset_id": {
                          "type": "string"
                        },
                        "side": {
                          "type": "string",
                          "enum": ["buy", "sell"]
                        },
                        "usd_amount": {
                          "type": "string"
                        },
                        "qty": {
                          "type": "string"
                        },
                        "price_at_execution": {
                          "type": "string"
                        },
                        "idempotency_key": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "asset_id",
                        "side",
                        "usd_amount",
                        "qty",
                        "price_at_execution",
                        "idempotency_key",
                        "created_at"
                      ],
                      "additionalProperties": false
                    },
                    "is_first_trade": {
                      "type": "boolean"
                    }
                  },
                  "required": ["trade", "is_first_trade"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "insufficient_cash",
                        "insufficient_qty",
                        "unknown_asset",
                        "invalid_amount"
                      ]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": ["price_unavailable"]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List the user's trades, newest first",
        "tags": ["trades"],
        "parameters": [
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trades": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "asset_id": {
                            "type": "string"
                          },
                          "side": {
                            "type": "string",
                            "enum": ["buy", "sell"]
                          },
                          "usd_amount": {
                            "type": "string"
                          },
                          "qty": {
                            "type": "string"
                          },
                          "price_at_execution": {
                            "type": "string"
                          },
                          "idempotency_key": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "asset_id",
                          "side",
                          "usd_amount",
                          "qty",
                          "price_at_execution",
                          "idempotency_key",
                          "created_at"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["trades"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}